Scan Rows into struct, fix #59
